CircularBuffer: Make _pool volatiile
Even if accesses to _pool can not be optimized because it's of T volatile *
(because of _head and _tail being volatile). Yet is should still be of
volatile T volatile *, because from the point of view of the thread the contents
of the dereferenced pointer can change without its interaction.